django_base/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyascii==0.3.2
asgiref==3.6.0
async-signals==0.1.5
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
blis==0.7.9
boto3==1.26.131
botocore==1.29.131
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
-e git+ssh://git@github.com/django/channels.git@fb1b86fbb345ac55c8b08b79044166965859bb59#egg=channels
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==39.0.2
cymem==2.0.7
daphne==4.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.7
django-allauth==0.53.1
django-appconf==1.0.5
django-bootstrap5==22.2
django-compressor==4.3.1
django-crispy-forms==2.0
django-extensions==3.2.1
django-filter==22.1
django-hosts==5.2
django-modelcluster==6.0
django-permissionedforms==0.1
django-sass-processor==1.2.2
django-simple-history==3.3.0
django-storages==1.13.2
django-summernote==0.8.20.0
django-taggit==3.1.0
django-treebeard==4.6.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
frozenlist==1.3.3
google-analytics==0.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
hyperframe==6.0.1
hyperlink==21.0.0
icalendar==5.0.4
idna==3.4
incremental==22.10.0
iso8601==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
l18n==2021.3
langcodes==3.3.0
libsass==0.22.0
MarkupSafe==2.1.2
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.4
openpyxl==3.1.2
packaging==23.0
pathspec==0.11.1
pathy==0.10.1
Pillow==9.5.0
platformdirs==3.1.1
preshed==3.0.8
priority==1.3.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
rcssmin==1.1.1
redis==4.5.4
redis-cli==0.1.2
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rjsmin==1.2.1
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
smart-open==6.3.0
sob==1.50.3
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
telepath==0.3
textblob==0.17.1
thinc==8.1.9
tomli==2.0.1
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
wagtail==4.2.2
wasabi==1.1.1
webencodings==0.5.1
Willow==1.4.1
yarl==1.8.2
zope.interface==6.0
